requirements.txt
alembic==1.11.1
altgraph==0.17.3
bcrypt==4.0.1
blinker==1.6.2
bottle==0.12.25
cffi==1.15.1
click==8.1.3
clr-loader==0.2.5
colorama==0.4.6
dnspython==2.3.0
email-validator==2.0.0.post2
et-xmlfile==1.1.0
Flask==2.3.2
Flask-Bcrypt==1.0.1
Flask-Login==0.6.2
Flask-Migrate==4.0.4
Flask-Reuploaded==1.3.0
Flask-SQLAlchemy==3.0.3
Flask-Uploads==0.2.1
Flask-WTF==1.1.1
greenlet==2.0.2
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.2
neat-python==0.92
openpyxl==3.1.2
pandas==2.0.2
pefile==2023.2.7
proxy-tools==0.1.0
pycparser==2.21
pyinstaller==5.11.0
pyinstaller-hooks-contrib==2023.3
python-dateutil==2.8.2
pythonnet==3.0.1
pytz==2023.3
pywebview==4.1
pywin32-ctypes==0.2.0
six==1.16.0
SQLAlchemy==2.0.15
typing_extensions==4.6.1
tzdata==2023.3
urllib3==2.0.2
Werkzeug==2.3.4
WTForms==3.0.1
